% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_methods.R
\name{.arrange_scatter_panels}
\alias{.arrange_scatter_panels}
\title{Arrange scatter panels in a grid}
\usage{
.arrange_scatter_panels(plot_list)
}
\description{
Arrange scatter panels in a grid
}
\keyword{internal}
