% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robust_nonparametric.R
\name{.generate_robust_xgb}
\alias{.generate_robust_xgb}
\title{Generate from robust XGBoost}
\usage{
.generate_robust_xgb(fit, X_new, n)
}
\description{
Generate from robust XGBoost
}
\keyword{internal}
