% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_registry.R
\name{.methods_for_type}
\alias{.methods_for_type}
\title{Method names that support a given variable type.}
\usage{
.methods_for_type(type)
}
\description{
Method names that support a given variable type.
}
\keyword{internal}
