% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_methods.R
\name{.plot_density}
\alias{.plot_density}
\title{Overlay density plot (used by both plot.robsynth and compare)}
\usage{
.plot_density(original, synth, var_types, vars)
}
\description{
Overlay density plot (used by both plot.robsynth and compare)
}
\keyword{internal}
