% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_methods.R
\name{.plot_scatter}
\alias{.plot_scatter}
\title{Pairwise scatter plots}
\usage{
.plot_scatter(
  original,
  synth,
  var_types,
  vars,
  layout = c("overlay", "facet"),
  alpha = 0.4
)
}
\description{
Pairwise scatter plots
}
\keyword{internal}
