% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robsynth.R
\name{.synth_marginal}
\alias{.synth_marginal}
\title{Marginal synthesis for a single variable}
\usage{
.synth_marginal(y, type, n, w = NULL)
}
\arguments{
\item{y}{vector of original values}

\item{type}{\code{"continuous"} or \code{"categorical"}}

\item{n}{number of synthetic values}
}
\value{
vector of length \code{n}
}
\description{
Marginal synthesis for a single variable
}
\keyword{internal}
