% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robust_conditional.R
\name{generate_robust_categorical}
\alias{generate_robust_categorical}
\title{Generate synthetic categorical values from a robust fit}
\usage{
generate_robust_categorical(fit, X_new)
}
\arguments{
\item{fit}{object of class \code{"robsynth_fit_cat"} as returned
by \code{\link{fit_robust_categorical}}}

\item{X_new}{data.frame of predictors (from synthetic data)}
}
\value{
factor vector of synthetic values
}
\description{
Draws from predicted class probabilities of the fitted weighted
multinomial model.
}
\examples{
\dontrun{
data(CrohnD, package = "robustbase")
fit <- fit_robust_categorical(CrohnD$treat,
                              CrohnD[, c("BMI", "age"), drop = FALSE])
y_new <- generate_robust_categorical(fit,
                              CrohnD[, c("BMI", "age"), drop = FALSE])
table(y_new)
}
}
\author{
Matthias Templ
}
