% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robsynth.R
\name{print.robsynth}
\alias{print.robsynth}
\title{Print method for robsynth objects}
\usage{
\method{print}{robsynth}(x, ...)
}
\arguments{
\item{x}{a \code{"robsynth"} object}

\item{...}{ignored}
}
\description{
Print method for robsynth objects
}
