% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchical.R
\name{print.robsynth_hh}
\alias{print.robsynth_hh}
\title{Print method for robsynth_hh objects}
\usage{
\method{print}{robsynth_hh}(x, ...)
}
\arguments{
\item{x}{a \code{"robsynth_hh"} object}

\item{...}{ignored}
}
\description{
Print method for robsynth_hh objects
}
