% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{print.ru_frontier}
\alias{print.ru_frontier}
\title{Print method for ru_frontier}
\usage{
\method{print}{ru_frontier}(x, ...)
}
\arguments{
\item{x}{a \code{"ru_frontier"} object}

\item{...}{ignored}
}
\description{
Print method for ru_frontier
}
