% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robSynth-package.R
\docType{package}
\name{robSynth-package}
\alias{robSynth}
\alias{robSynth-package}
\title{robSynth: Robust Synthetic Data Generation}
\description{
Robust synthetic data generation methods that protect data quality
from contaminated training data.  Replaces the conditional models in
sequential synthesis with MM-estimators (continuous) and weighted
logistic regression (categorical).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/statistikat/robSynth}
  \item Report bugs at \url{https://github.com/statistikat/robSynth/issues}
}

}
\author{
\strong{Maintainer}: Matthias Templ \email{matthias.templ@fhnw.ch} (\href{https://orcid.org/0000-0002-8638-5276}{ORCID})

}
\keyword{internal}
