% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{strip_diagnostics}
\alias{strip_diagnostics}
\title{Strip diagnostics from a robsynth result}
\usage{
strip_diagnostics(x)
}
\arguments{
\item{x}{an object of class \code{"robsynth"}}
}
\value{
the same object with \code{contWeights} set to \code{NULL}
}
\description{
Removes the \code{contWeights} component from a robsynth result
to prevent disclosure of contamination flags. This should be
called before releasing synthetic data.
}
\author{
Matthias Templ
}
