% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robsynth.R
\name{summary.robsynth}
\alias{summary.robsynth}
\title{Summary method for robsynth objects}
\usage{
\method{summary}{robsynth}(object, ...)
}
\arguments{
\item{object}{a \code{"robsynth"} object}

\item{...}{ignored}
}
\description{
Summary method for robsynth objects
}
