% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robsynth.R
\name{synth_glm}
\alias{synth_glm}
\title{Fit a GLM on synthetic data}
\usage{
synth_glm(formula, object, family = stats::gaussian(), ...)
}
\arguments{
\item{formula}{a model formula}

\item{object}{a \code{"robsynth"} object}

\item{family}{a family object (e.g., \code{binomial()})}

\item{...}{additional arguments passed to \code{glm}}
}
\value{
same structure as \code{\link{synth_lm}}
}
\description{
Fit a GLM on synthetic data
}
